small refactoring database names and connections
