+++ src/auth/strategies/jwt.strategy.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@nestjs/common';
 import { PassportStrategy } from '@nestjs/passport';
 import { Strategy, ExtractJwt } from 'passport-jwt';
 import { ConfigService } from '@nestjs/config';
-import { PrismaService } from '../prisma.service';
+import { PrismaService } from '../../prisma.service';
 import { User } from '@prisma/client';
 
 @Injectable()
